pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "irtk"
authors = [{name = "Guangyan Cai", email = "gcai3@uci.edu"}]
readme = "README.md"
dynamic = ["version", "description"]
requires-python = ">=3.8"
dependencies = [
    'imageio >= 2.27.0',
    'imageio_ffmpeg',
    'numpy',
    'gin-config',
    'gpytoolbox',
    'trimesh',
    'chamferdist=1.0.0',
]

[project.urls]
Home = "https://github.com/uci-rendering/inv-render-toolkit"

[tool.flit.sdist]
exclude = [
    "examples/",
    "project_template/"
]